# CarCompany version 2 (ASP Assignment 2)/CarCompany/CarCompany/Models/Skill.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace CarCompany.Models
{
    public class Skill
    {
        public long SkillId { get; set; }
        public string Name { get; set; }
        public string Description { get; set; }
        public Company CompanyID { get; set; }
    }
}